Matching + Logout
Added a user_id column to our requests table and added a matching implementation that appears on the home screen based on all user data. So now, if a user makes a request (i.e. they have section 5 of CS112 and want section 7, and someone else has section 7 and wants section 5), the home screen will display this match for the user.

I also implemented a logout button so this way it is easier to test. Instead of having to shut down the program to log into a new user, you can simply log out and log back in and test by making new trade requests.

The next step is to polish the frontend for the showcase.